now getting clustered groups of students. However, as the algo stands these groups are not distinct...i.e devoid of intersections. We need to think of how to consolidate.....we will likely merge "interesting clusters" with common elems, even if all the students there are not necessarily beyond the threshold.

+++ src/model/StudentPair.java
@@ -4,11 +4,11 @@ public class StudentPair {
 	
 	Student left;
 	Student right;
-	float rawScore;
-	float relativeScore;
+	double rawScore;
+	double relativeScore;
 	String id;
 	
-	public StudentPair(Student left, Student right, float rawScore, float relativeScore){
+	public StudentPair(Student left, Student right, double rawScore, double relativeScore){
 		this.left=left;
 		this.right=right;
 		this.rawScore=rawScore;
@@ -24,11 +24,11 @@ public class StudentPair {
 		return right;
 	}
 
-	public float getRawScore() {
+	public double getRawScore() {
 		return rawScore;
 	}
 
-	public float getRelativeScore() {
+	public double getRelativeScore() {
 		return relativeScore;
 	}
